packages/qfull-chemistry/pyproject.toml
# qfull-chemistry: chemistry domain plugin for the QCompass workspace.
#
# Implements the qcompass_core.Simulation protocol so callers route
# molecular-electronic-structure problems through three execution
# paths: classical (PySCF + block2), Qiskit SQD, and Dice SHCI.
# Heavy SDKs are optional extras so the base install stays lean.
#
# Boundary: this package MUST NOT import from `ucgle_f1` or any other
# `qfull_*` sibling. CI enforces with both a grep guard and a static
# pytest check (tests/test_boundary.py).

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "qfull-chemistry"
version = "0.1.0"
description = "QFull-Chemistry — quantum-chemistry domain plugin for QCompass."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "QCompass authors" }]
dependencies = [
  "pydantic>=2.12",
  "numpy>=1.26",
  "scipy>=1.14",
  "h5py>=3.12",
  "pyyaml>=6",
  "qcompass-core>=0.1,<0.2",
  "openfermion>=1.7",
  "openfermionpyscf>=0.5",
]

[project.optional-dependencies]
chem = [
  "pyscf>=2.7",
  "block2>=0.5.3",
]
sqd = [
  "qiskit>=2.2,<3",
  "qiskit-aer>=0.17",
  "qiskit-addon-sqd",
  "qiskit-addon-utils>=0.4",
  "ffsim>=0.0.65",
]
dice = [
  "qiskit-addon-dice-solver; sys_platform == 'linux'",
]
ibm = [
  "qiskit-ibm-runtime",
]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "pytest-regressions>=2.5",
  "hypothesis>=6.120",
  "mypy>=1.13",
  "ruff>=0.8",
]

[project.entry-points."qcompass.domains"]
chemistry = "qfull_chem.sim:ChemistrySimulation"

[tool.hatch.build.targets.wheel]
packages = ["src/qfull_chem"]

[tool.hatch.build.targets.wheel.force-include]
"src/qfull_chem/instances" = "qfull_chem/instances"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
  "pyscf.*", "block2.*", "pyblock2.*",
  "qiskit.*", "qiskit_aer.*", "qiskit_ibm_runtime.*",
  "qiskit_addon_sqd.*", "qiskit_addon_dice_solver.*",
  "qiskit_addon_utils.*",
  "openfermion.*", "openfermionpyscf.*",
  "ffsim.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests", "src/qfull_chem/audit"]
asyncio_mode = "auto"
markers = [
  "slow: long-running test",
  "s_audit: per-domain S-chem-* audit",
  "needs_pyscf: requires pyscf",
  "needs_block2: requires pyblock2",
  "needs_sqd: requires qiskit + qiskit-addon-sqd",
  "needs_dice: requires qiskit-addon-dice-solver (Linux only)",
  "dice_only: test that exercises the Dice path; skipped on non-Linux",
]
